set up form fncs to return active layer values
