+++ maps/full-template/index.html
@@ -42,7 +42,7 @@
                 
                 <span class="nav-vr">|</span>
 
-                <button id="modal-toggle" class="nav-btn nav-links">about</button>
+                <button id="modal-toggle" class="nav-btn">about</button>
             </div>
         </nav>
 
@@ -87,12 +87,12 @@
 
                             <label class="label-input">
                                 layer 3
-                                <input type="checkbox" value="4">
+                                <input type="checkbox" value="3">
                             </label>
         
                             <label class="label-input">
                                 layer 4
-                                <input type="checkbox" value="5">
+                                <input type="checkbox" value="4">
                             </label>
                         </fieldset>
                     </form>
